reverting mesos-go paths to test PR
