+++ scheduler/riak_node_state.go
@@ -1,10 +1,30 @@
 package scheduler
 
 import (
+	mesos "github.com/mesos/mesos-go/mesosproto"
+	util "github.com/mesos/mesos-go/mesosutil"
 	"github.com/basho-labs/riak-mesos/scheduler/process_state"
 )
 
 // Getters
+func (frn *FrameworkRiakNode) HasRequestedReservation() bool {
+	if frn.LastOfferUsed == nil {
+		return false
+	}
+
+	for _, resource := range util.FilterResources(frn.LastOfferUsed.Resources, func(res *mesos.Resource) bool { return res.GetName() == "disk" && res.Reservation != nil }) {
+		if resource.Disk != nil && *resource.Disk.Persistence.Id == frn.PersistenceID() {
+			return true
+		}
+	}
+
+	if frn.CurrentState == process_state.ReservationRequested {
+		return true
+	}
+
+	return false
+}
+
 func (frn *FrameworkRiakNode) CanBeScheduled() bool {
 	switch frn.DestinationState {
 	case process_state.Started:
