+++ scheduler/scheduler.go
@@ -7,10 +7,10 @@ import (
 	"golang.org/x/net/context"
 
 	log "github.com/Sirupsen/logrus"
-	auth "github.com/basho-labs/mesos-go/auth"
-	sasl "github.com/basho-labs/mesos-go/auth/sasl"
-	mesos "github.com/basho-labs/mesos-go/mesosproto"
-	sched "github.com/basho-labs/mesos-go/scheduler"
+	auth "github.com/mesos/mesos-go/auth"
+	sasl "github.com/mesos/mesos-go/auth/sasl"
+	mesos "github.com/mesos/mesos-go/mesosproto"
+	sched "github.com/mesos/mesos-go/scheduler"
 	"github.com/basho-labs/riak-mesos/cepmd/cepm"
 	metamgr "github.com/basho-labs/riak-mesos/metadata_manager"
 	"github.com/golang/protobuf/proto"
