+++ setup-env.sh
@@ -26,13 +26,13 @@ cd $GOPATH/src/github.com/basho-labs/riak-mesos && godep restore
 REAL_GOPATH=$GOPATH
 gvm use go1.4
 export GOPATH=$REAL_GOPATH
-rm -rf $GOPATH/src/github.com/basho-labs/mesos-go
-git clone https://github.com/basho-labs/mesos-go.git $GOPATH/src/github.com/basho-labs/mesos-go
+rm -rf $GOPATH/src/github.com/mesos/mesos-go
+git clone https://github.com/mesos/mesos-go.git $GOPATH/src/github.com/mesos/mesos-go
 go get github.com/golang/protobuf/proto
 go get github.com/gogo/protobuf/proto
 go get github.com/gogo/protobuf/protoc-gen-gogo
 go get github.com/gogo/protobuf/protoc-gen-gogofast
-cd $GOPATH/src/github.com/basho-labs/mesos-go/mesosproto && \
+cd $GOPATH/src/github.com/mesos/mesos-go/mesosproto && \
   protoc --proto_path=${GOPATH}/src:${GOPATH}/src/github.com/gogo/protobuf/protobuf:. --gogofast_out=. *.proto
 gvm use go1.5
 export GOPATH=$REAL_GOPATH
